pyproject.toml
[project]
name = "trafficdetector"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "asgi-correlation-id>=4.3.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.0",
    "celery-types==0.23.0",
    "fastapi[standard]>=0.123.4",
    "google-cloud-bigquery>=3.13.0",
    "gradio>=6.0.2",
    "httpx>=0.28.1",
    "opencv-python-headless>=4.12.0.88",
    "pre-commit>=4.5.0",
    "psycopg[binary]>=3.3.1",
    "pydantic-settings>=2.12.0",
    "pyright>=1.1.407",
    "pytest-asyncio>=1.3.0",
    "python-logging-loki>=0.3.1",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "ruff>=0.14.7",
    "sqlmodel>=0.0.27",
    "structlog>=25.5.0",
    "supervision>=0.27.0",
    "ultralytics>=8.3.234",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "ruff>=0.14.7",
    "pyright>=1.1.370"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "build",
    "dist",
]

line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend"]

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "Linux"
typeCheckingMode = "basic"
include = ["backend", "frontend"]
exclude = ["backend/tests", "frontend/tests", "**/__pycache__"]
reportMissingTypeStubs = "none"
reportUnknownVariableType = "none"
reportUnknownMemberType = "none"
reportGeneralTypeIssues = "warning"
reportUnknownArgumentType = "none"
reportUnusedImport = false
reportPrivateImportUsage = "none"
reportAttributeAccessIssue = "none"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["backend/tests", "frontend/tests"]

addopts = "--asyncio-mode=auto"

asyncio_default_fixture_loop_scope = "function"

filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestRemovedIn9Warning",
    "ignore::RuntimeWarning:asyncio"
]
